0Auth/django/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
requests==2.31.0
cryptography==41.0.7
python-decouple==3.8
Pillow>=10.0.0
dj-database-url==2.1.0
django-allauth==0.57.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
PyMySQL==1.1.2